split render.rs
