+++ src/ui/render/table.rs
@@ -1,0 +1,24 @@
+use ratatui::{
+    layout::{Constraint, Rect},
+    style::{Color, Modifier, Style, Stylize},
+    widgets::{Block, BorderType, Row, Table},
+    Frame,
+};
+
+use crate::ui::app::App;
+
+pub fn render(app: &mut App, frame: &mut Frame, area: Rect) {
+    let widths = [Constraint::Length(6), Constraint::Percentage(100), Constraint::Length(10), Constraint::Length(10)];
+    let table_data = app.table.clone();
+    let table = Table::new(table_data.to_rows(app.args.no_icons), widths)
+        .column_spacing(1)
+        .header(
+            Row::new(vec!["", "Path", "LastMod", "Size"])
+                .style(Style::default().bg(Color::Cyan).add_modifier(Modifier::BOLD)),
+        )
+        .block(Block::bordered().border_type(BorderType::Rounded))
+        .highlight_style(Style::default().reversed())
+        .highlight_symbol(" ");
+
+    frame.render_stateful_widget(table, area, &mut app.table.state);
+}
